// Client/bindings/FieldEventTypeField.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FieldEventValueKind } from "./FieldEventValueKind";

export type FieldEventTypeField = {
  id: number;
  name: string;
  value_kind: FieldEventValueKind;
  unit: string | null;
};